Logout update

+++ app/Http/Controllers/Api/AuthController.php
@@ -18,14 +18,12 @@ class AuthController extends Controller
         $headers = [
             'Access-Control-Allow-Origin' => '*',
         ];
-        Log::info('reached signup method');
         $data = $request->validated();
         if ($data['type'] === 'student') {
             $user = Student::create([
                 'name' => $data['name'],
                 'email' => $data['email'],
                 'password' => bcrypt($data['password']),
-
             ]);
         } else if ($data['type'] === 'instructor') {
             $user = Instructor::create([
@@ -45,26 +43,44 @@ class AuthController extends Controller
     public function login(LoginRequest $request)
     {
         $credentials = $request->validated();
-        if (!Auth::attempt($credentials)) {
+        $userType = $request->input('type');
+
+        if ($userType === 'student') {
+            $guard = 'student';
+        } else if ($userType === 'instructor') {
+            $guard = 'instructor';
+        } else {
             return response()->json([
-                'message' => 'Invalid login credentials'
-            ], 401);
+                'message' => 'Invalid user type'
+            ], 422);
         }
 
-        $user = Auth::user();
-        $token = $user->createToken('auth_token')->plainTextToken;
-        return response(compact('user', 'token'));
-
-
+        if (Auth::guard($guard)->attempt($credentials)) {
+            $user = Auth::guard($guard)->user();
+            $token = $user->createToken('auth_token')->plainTextToken;
+            return response(compact('user', 'token'));
+        } else {
+            return response()->json([
+                'message' => "Invalid $userType login credentials"
+            ], 401);
+        }
     }
 
 
     public function logout(Request $request)
     {
         $user = $request->user();
-        $user->currentAccessToken()->delete();
+
+        if ($user) {
+            $user->currentAccessToken()->delete();
+            return response()->json([
+                'message' => 'Logged out'
+            ]);
+        }
+
+        // If the user is not authenticated, return an appropriate response
         return response()->json([
-            'message' => 'Logged out'
-        ]);
+            'message' => 'User not authenticated'
+        ], 401);
     }
 }
